Update!
fBaseXtensions
Added RemovalTypes enum and property to CacheObject and ObjectRemovedArgs class
Bot stop event will reset CharacterControl variables.
Added Line of Sight control settings for maximum and minimum ranges.
Character Control hero info class no longer uses time played in equal method.

fItemPlugin minimum blood shard setting is now 25.

FunkyKeyBot
Replaced FunkyWhileActorPresent to use FunkyWhile and custom condition.

+++ fBaseXtensions/Settings/_LOSMovementSetting.cs
@@ -14,8 +14,8 @@
 		public bool AllowEventSwitches { get; set; }
 		public int MaximumRange { get; set; }
 
-		public float MiniumRangeObjects { get; set; }
-		public float MinimumRangeMarkers { get; set; }
+		public int MiniumRangeObjects { get; set; }
+        public int MinimumRangeMarkers { get; set; }
 
         public SettingLOSMovement()
         {
